requirements.txt
discord.py
# openai  <-- Comentado para que no se instale
groq
python-dotenv
